move *Tp3 to tinkerpop subpackage
